Move athenz api to separate package
